bot/requirements.txt
discord.py[voice]
asyncpg
tortoise-orm
python-dotenv
logging
pathlib